// src/types/enums/ClientSocketEvents.ts
export enum ClientSocketEvents {
	SetUuid = "setuuid",
	JoinRoom = "joinroom",
	Disconnect = "disconnect",
	Error = "error",
	SyncAllClients = "syncallclients",
	AddClient = "addclient",
	RemoveClient = "removeclient",
	SetPositionOf = "setpositionof",
	SetVentOf = "setventof",
	SetNameOf = "setnameof",
	SetColorOf = "setcolorof",
	SetHatOf = "sethatof",
	SetSkinOf = "setskinof",
	SetHost = "sethost",
	SetOptions = "setoptions",
	SetSettings = "setsettings",
	SetGameState = "setgamestate",
	SetGameFlags = "setgameflags",
	SetFlagsOf = "setflagsof",
}